sunxi: Explicitly cast u32 pointer conversions
Some parts of the sunxi code cast explicitly between u32 values and pointers.
This is not a problem in practice, because all 64bit SoCs today only use the
lower 32 bits for their phyical address space. But we need to make sure that
the compiler is sure this is not an accident as well.

Acked-by: Hans de Goede <hdegoede@redhat.com>

+++ arch/arm/mach-sunxi/cpu_info.c
@@ -105,7 +105,7 @@ int sunxi_get_sid(unsigned int *sid)
 	int i;
 
 	for (i = 0; i< 4; i++)
-		sid[i] = readl(SUNXI_SID_BASE + 4 * i);
+		sid[i] = readl((ulong)SUNXI_SID_BASE + 4 * i);
 
 	return 0;
 #else
